+++ src/lib/components/Row.svelte
@@ -1,8 +1,15 @@
 <script lang="ts">
-	let { children } = $props();
+	import type { Snippet } from 'svelte';
+
+	interface Props {
+		children: Snippet;
+		baseline?: boolean;
+	}
+
+	let { children, baseline = false }: Props = $props();
 </script>
 
-<div class="row">{@render children()}</div>
+<div class="row" class:baseline>{@render children()}</div>
 
 <style>
 	.row {
@@ -12,5 +19,9 @@
 		gap: var(--spacing);
 		row-gap: var(--spacing);
 		align-items: center;
+	}
+
+	.baseline {
+		align-items: baseline;
 	}
 </style>